fix test

+++ packages/server/customer-os-api/graph/resolver/service_line_item.resolvers_it_test.go
@@ -36,7 +36,7 @@ func TestMutationResolver_ServiceLineItemCreate(t *testing.T) {
 			require.Equal(t, "Service Line Item 1", serviceLineItem.Name)
 			require.Equal(t, servicelineitempb.BilledType_MONTHLY_BILLED, serviceLineItem.Billed)
 			require.Equal(t, int64(2), serviceLineItem.Quantity)
-			require.Equal(t, float32(30), serviceLineItem.Price)
+			require.Equal(t, float64(30), serviceLineItem.Price)
 
 			calledCreateServiceLineItem = true
 			neo4jt.CreateServiceLineItemForContract(ctx, driver, tenantName, contractId, entity.ServiceLineItemEntity{
@@ -88,7 +88,7 @@ func TestMutationResolver_ServiceLineItemUpdate(t *testing.T) {
 			require.Equal(t, "Service Line Item 1", serviceLineItem.Name)
 			require.Equal(t, servicelineitempb.BilledType_MONTHLY_BILLED, serviceLineItem.Billed)
 			require.Equal(t, int64(2), serviceLineItem.Quantity)
-			require.Equal(t, float32(30), serviceLineItem.Price)
+			require.Equal(t, float64(30), serviceLineItem.Price)
 			require.Equal(t, "test comments", serviceLineItem.Comments)
 			require.Equal(t, true, serviceLineItem.IsRetroactiveCorrection)
 			calledUpdateServiceLineItem = true
